+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Show annotation method default values in the structure view
 - Show inner class details in the structure view
 - Show try-catch type annotations in the structure view
+- Show type annotation target metadata in the structure view
 
 ### Changed